refactor to address static code analysis tool remarks
