+++ mesonbuild/cmake/generator.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from .. import mesonlib
 from .. import mlog
 from ..mesonlib import OptionKey
-from .tracetargets import get_config_declined_property
+from .common import get_config_declined_property
 import typing as T
 
 if T.TYPE_CHECKING:
